+++ src/resolvers/import-queue-resolver.ts
@@ -3,12 +3,12 @@ import Resolver from '@forge/resolver';
 import { storage } from '@forge/api';
 import { backOff, IBackOffOptions } from 'exponential-backoff';
 
-import { createComponent, updateComponent } from './client/compass';
-import { STORAGE_KEYS, BACK_OFF, IMPORT_LABEL } from './constants';
-import { appendLink } from './utils/append-link';
+import { createComponent, updateComponent } from '../client/compass';
+import { STORAGE_KEYS, BACK_OFF, IMPORT_LABEL } from '../constants';
+import { appendLink } from '../utils/append-link';
 import { ImportableProject } from './resolverTypes';
-import { sleep } from './utils/time-utils';
-import { createMRWithCompassYML } from './services/create-mr-with-compass-yml';
+import { sleep } from '../utils/time-utils';
+import { createMRWithCompassYML } from '../services/create-mr-with-compass-yml';
 
 const backOffConfig: Partial<IBackOffOptions> = {
   startingDelay: BACK_OFF.startingDelay,
